Rename SettingsActivity and optimize text
